Extract CasperMessage and introduce toCasperMessage(Packet) method
